CI: Set file owner for the whole working tree including submodules
This is important for files which get copied from sources/meta-*
directly. E.g. this happens for some license files. When the image is
completed bitbake tries to set the atime (access time). It seems that on
Linux only owners of files are allowed to update atime of files. Since
we have to run bitbake as normal user, this fixup is needed.

Signed-off-by: Moritz Scheuren <moritz.scheuren@systec-electronic.com>